Moved functions that aren't critical for build
